+++ statsd/statsd.go
@@ -57,15 +57,31 @@ const DefaultMaxAgentPayloadSize = 8192
 
 /*
 UnixAddressPrefix holds the prefix to use to enable Unix Domain Socket
-traffic instead of UDP.
+traffic instead of UDP. The type of the socket will be guessed.
 */
 const UnixAddressPrefix = "unix://"
+
+/*
+UnixDatagramAddressPrefix holds the prefix to use to enable Unix Domain Socket
+datagram traffic instead of UDP.
+*/
+const UnixAddressDatagramPrefix = "unixgram://"
+
+/*
+UnixAddressStreamPrefix holds the prefix to use to enable Unix Domain Socket
+stream traffic instead of UDP.
+*/
+const UnixAddressStreamPrefix = "unixstream://"
 
 /*
 WindowsPipeAddressPrefix holds the prefix to use to enable Windows Named Pipes
 traffic instead of UDP.
 */
 const WindowsPipeAddressPrefix = `\\.\pipe\`
+
+var (
+	AddressPrefixes = []string{UnixAddressPrefix, UnixAddressDatagramPrefix, UnixAddressStreamPrefix, WindowsPipeAddressPrefix}
+)
 
 const (
 	agentHostEnvVarName = "DD_AGENT_HOST"
@@ -240,6 +256,8 @@ type ClientInterface interface {
 	GetTelemetry() Telemetry
 }
 
+type ErrorHandler func(error)
+
 // A Client is a handle for sending messages to dogstatsd.  It is safe to
 // use one Client from multiple goroutines simultaneously.
 type Client struct {
@@ -248,21 +266,23 @@ type Client struct {
 	// namespace to prepend to all statsd calls
 	namespace string
 	// tags are global tags to be added to every statsd call
-	tags            []string
-	flushTime       time.Duration
-	telemetry       *statsdTelemetry
-	telemetryClient *telemetryClient
-	stop            chan struct{}
-	wg              sync.WaitGroup
-	workers         []*worker
-	closerLock      sync.Mutex
-	workersMode     receivingMode
-	aggregatorMode  receivingMode
-	agg             *aggregator
-	aggExtended     *aggregator
-	options         []Option
-	addrOption      string
-	isClosed        bool
+	tags                  []string
+	flushTime             time.Duration
+	telemetry             *statsdTelemetry
+	telemetryClient       *telemetryClient
+	stop                  chan struct{}
+	wg                    sync.WaitGroup
+	workers               []*worker
+	closerLock            sync.Mutex
+	workersMode           receivingMode
+	aggregatorMode        receivingMode
+	agg                   *aggregator
+	aggExtended           *aggregator
+	options               []Option
+	addrOption            string
+	isClosed              bool
+	errorOnBlockedChannel bool
+	errorHandler          ErrorHandler
 }
 
 // statsdTelemetry contains telemetry metrics about the client
@@ -301,14 +321,19 @@ func resolveAddr(addr string) string {
 		return ""
 	}
 
-	if !strings.HasPrefix(addr, WindowsPipeAddressPrefix) && !strings.HasPrefix(addr, UnixAddressPrefix) {
-		if !strings.Contains(addr, ":") {
-			if envPort != "" {
-				addr = fmt.Sprintf("%s:%s", addr, envPort)
-			} else {
-				addr = fmt.Sprintf("%s:%s", addr, defaultUDPPort)
-			}
+	for _, prefix := range AddressPrefixes {
+		if strings.HasPrefix(addr, prefix) {
+			return addr
 		}
+	}
+	// TODO: How does this work for IPv6?
+	if strings.Contains(addr, ":") {
+		return addr
+	}
+	if envPort != "" {
+		addr = fmt.Sprintf("%s:%s", addr, envPort)
+	} else {
+		addr = fmt.Sprintf("%s:%s", addr, defaultUDPPort)
 	}
 	return addr
 }
@@ -349,7 +374,13 @@ func createWriter(addr string, writeTimeout time.Duration, udpAddrRefreshRate ti
 		w, err := newWindowsPipeWriter(addr, writeTimeout)
 		return w, writerWindowsPipe, err
 	case strings.HasPrefix(addr, UnixAddressPrefix):
-		w, err := newUDSWriter(addr[len(UnixAddressPrefix):], writeTimeout)
+		w, err := newUDSWriter(addr[len(UnixAddressPrefix):], writeTimeout, "")
+		return w, writerNameUDS, err
+	case strings.HasPrefix(addr, UnixAddressDatagramPrefix):
+		w, err := newUDSWriter(addr[len(UnixAddressDatagramPrefix):], writeTimeout, "unixgram")
+		return w, writerNameUDS, err
+	case strings.HasPrefix(addr, UnixAddressStreamPrefix):
+		w, err := newUDSWriter(addr[len(UnixAddressStreamPrefix):], writeTimeout, "unix")
 		return w, writerNameUDS, err
 	default:
 		w, err := newUDPWriter(addr, writeTimeout, udpAddrRefreshRate)
@@ -403,9 +434,11 @@ func CloneWithExtraOptions(c *Client, options ...Option) (*Client, error) {
 
 func newWithWriter(w io.WriteCloser, o *Options, writerName string) (*Client, error) {
 	c := Client{
-		namespace: o.namespace,
-		tags:      o.tags,
-		telemetry: &statsdTelemetry{},
+		namespace:             o.namespace,
+		tags:                  o.tags,
+		telemetry:             &statsdTelemetry{},
+		errorOnBlockedChannel: o.channelModeErrorsWhenFull,
+		errorHandler:          o.errorHandler,
 	}
 
 	hasEntityID := false
@@ -446,7 +479,7 @@ func newWithWriter(w io.WriteCloser, o *Options, writerName string) (*Client, er
 	}
 
 	bufferPool := newBufferPool(o.bufferPoolSize, o.maxBytesPerPayload, o.maxMessagesPerPayload)
-	c.sender = newSender(w, o.senderQueueSize, bufferPool)
+	c.sender = newSender(w, o.senderQueueSize, bufferPool, o.errorHandler)
 	c.aggregatorMode = o.receiveMode
 
 	c.workersMode = o.receiveMode
@@ -458,8 +491,8 @@ func newWithWriter(w io.WriteCloser, o *Options, writerName string) (*Client, er
 		c.workersMode = mutexMode
 	}
 
-	if o.aggregation || o.extendedAggregation {
-		c.agg = newAggregator(&c)
+	if o.aggregation || o.extendedAggregation || o.maxBufferedSamplesPerContext > 0 {
+		c.agg = newAggregator(&c, int64(o.maxBufferedSamplesPerContext))
 		c.agg.start(o.aggregationFlushInterval)
 
 		if o.extendedAggregation {
@@ -567,6 +600,16 @@ func (c *Client) GetTelemetry() Telemetry {
 	return c.telemetryClient.getTelemetry()
 }
 
+type ErrorInputChannelFull struct {
+	Metric      metric
+	ChannelSize int
+	Msg         string
+}
+
+func (e ErrorInputChannelFull) Error() string {
+	return e.Msg
+}
+
 func (c *Client) send(m metric) error {
 	h := hashString32(m.name)
 	worker := c.workers[h%uint32(len(c.workers))]
@@ -576,6 +619,13 @@ func (c *Client) send(m metric) error {
 		case worker.inputMetrics <- m:
 		default:
 			atomic.AddUint64(&c.telemetry.totalDroppedOnReceive, 1)
+			err := &ErrorInputChannelFull{m, len(worker.inputMetrics), "Worker input channel full"}
+			if c.errorHandler != nil {
+				c.errorHandler(err)
+			}
+			if c.errorOnBlockedChannel {
+				return err
+			}
 		}
 		return nil
 	}
@@ -594,10 +644,18 @@ func (c *Client) sendBlocking(m metric) error {
 
 func (c *Client) sendToAggregator(mType metricType, name string, value float64, tags []string, rate float64, f bufferedMetricSampleFunc) error {
 	if c.aggregatorMode == channelMode {
+		m := metric{metricType: mType, name: name, fvalue: value, tags: tags, rate: rate}
 		select {
-		case c.aggExtended.inputMetrics <- metric{metricType: mType, name: name, fvalue: value, tags: tags, rate: rate}:
+		case c.aggExtended.inputMetrics <- m:
 		default:
 			atomic.AddUint64(&c.telemetry.totalDroppedOnReceive, 1)
+			err := &ErrorInputChannelFull{m, len(c.aggExtended.inputMetrics), "Aggregator input channel full"}
+			if c.errorHandler != nil {
+				c.errorHandler(err)
+			}
+			if c.errorOnBlockedChannel {
+				return err
+			}
 		}
 		return nil
 	}