+++ statsd/uds_test.go
@@ -1,9 +1,11 @@
+//go:build !windows
 // +build !windows
 
 package statsd
 
 import (
-	"fmt"
+	"encoding/binary"
+	"golang.org/x/net/nettest"
 	"math/rand"
 	"net"
 	"os"
@@ -19,13 +21,20 @@ func init() {
 }
 
 func TestNewUDSWriter(t *testing.T) {
-	w, err := newUDSWriter("/tmp/test.socket", 100*time.Millisecond)
+	w, err := newUDSWriter("/tmp/test.socket", 100*time.Millisecond, "")
+	assert.NotNil(t, w)
+	assert.NoError(t, err)
+	w, err = newUDSWriter("/tmp/test.socket", 100*time.Millisecond, "unix")
+	assert.NotNil(t, w)
+	assert.NoError(t, err)
+	w, err = newUDSWriter("/tmp/test.socket", 100*time.Millisecond, "unixgram")
 	assert.NotNil(t, w)
 	assert.NoError(t, err)
 }
 
-func TestUDSWrite(t *testing.T) {
-	socketPath := fmt.Sprintf("/tmp/dsd_%d.socket", rand.Int())
+func TestUDSDatagramWrite(t *testing.T) {
+	socketPath, err := nettest.LocalPath()
+	require.NoError(t, err)
 	defer os.Remove(socketPath)
 
 	address, err := net.ResolveUnixAddr("unixgram", socketPath)
@@ -35,15 +44,16 @@ func TestUDSWrite(t *testing.T) {
 	err = os.Chmod(socketPath, 0722)
 	require.NoError(t, err)
 
-	w, err := newUDSWriter(socketPath, 100*time.Millisecond)
+	w, err := newUDSWriter(socketPath, 100*time.Millisecond, "")
 	require.Nil(t, err)
 	require.NotNil(t, w)
 
 	// test 2 Write: the first one should setup the connection
 	for i := 0; i < 2; i++ {
-		n, err := w.Write([]byte("some data"))
+		msg := []byte("some data")
+		n, err := w.Write(msg)
 		require.NoError(t, err)
-		assert.Equal(t, 9, n)
+		assert.Equal(t, len(msg), n)
 
 		buffer := make([]byte, 100)
 		n, err = conn.Read(buffer)
@@ -52,8 +62,9 @@ func TestUDSWrite(t *testing.T) {
 	}
 }
 
-func TestUDSWriteUnsetConnection(t *testing.T) {
-	socketPath := fmt.Sprintf("/tmp/dsd_%d.socket", rand.Int())
+func TestUDSDatagramWriteUnsetConnection(t *testing.T) {
+	socketPath, err := nettest.LocalPath()
+	require.NoError(t, err)
 	defer os.Remove(socketPath)
 
 	address, err := net.ResolveUnixAddr("unixgram", socketPath)
@@ -63,15 +74,16 @@ func TestUDSWriteUnsetConnection(t *testing.T) {
 	err = os.Chmod(socketPath, 0722)
 	require.NoError(t, err)
 
-	w, err := newUDSWriter(socketPath, 100*time.Millisecond)
+	w, err := newUDSWriter(socketPath, 100*time.Millisecond, "")
 	require.Nil(t, err)
 	require.NotNil(t, w)
 
 	// test 2 Write: the first one should setup the connection
 	for i := 0; i < 2; i++ {
-		n, err := w.Write([]byte("some data"))
+		msg := []byte("some data")
+		n, err := w.Write(msg)
 		require.NoError(t, err)
-		assert.Equal(t, 9, n)
+		assert.Equal(t, len(msg), n)
 
 		buffer := make([]byte, 100)
 		n, err = conn.Read(buffer)
@@ -80,5 +92,91 @@ func TestUDSWriteUnsetConnection(t *testing.T) {
 
 		// Unset connection for the next Read
 		w.unsetConnection()
+	}
+}
+
+func TestUDSStreamWrite(t *testing.T) {
+	socketPath, err := nettest.LocalPath()
+	require.NoError(t, err)
+	defer os.Remove(socketPath)
+
+	address, err := net.ResolveUnixAddr("unix", socketPath)
+	require.NoError(t, err)
+	listener, err := net.ListenUnix("unix", address)
+	require.NoError(t, err)
+	err = os.Chmod(socketPath, 0722)
+	require.NoError(t, err)
+
+	w, err := newUDSWriter(socketPath, 100*time.Millisecond, "")
+	require.Nil(t, err)
+	require.NotNil(t, w)
+
+	var conn net.Conn
+
+	// test 2 Write: the first one should setup the connection
+	for i := 0; i < 2; i++ {
+		msg := []byte("some data")
+		n, err := w.Write(msg)
+		require.NoError(t, err)
+		assert.Equal(t, len(msg), n)
+
+		if conn == nil {
+			conn, err = listener.Accept()
+			require.NoError(t, err)
+		}
+
+		var l uint32
+		binary.Read(conn, binary.LittleEndian, &l)
+		assert.Equal(t, uint32(len(msg)), l)
+
+		buffer := make([]byte, 100)
+		n, err = conn.Read(buffer)
+		require.NoError(t, err)
+		assert.Equal(t, "some data", string(buffer[:n]))
+	}
+}
+
+func TestUDSStreamWriteUnsetConnection(t *testing.T) {
+	socketPath, err := nettest.LocalPath()
+	require.NoError(t, err)
+	defer os.Remove(socketPath)
+
+	address, err := net.ResolveUnixAddr("unix", socketPath)
+	require.NoError(t, err)
+	listener, err := net.ListenUnix("unix", address)
+	require.NoError(t, err)
+	err = os.Chmod(socketPath, 0722)
+	require.NoError(t, err)
+
+	w, err := newUDSWriter(socketPath, 100*time.Millisecond, "")
+	require.Nil(t, err)
+	require.NotNil(t, w)
+
+	var conn net.Conn
+
+	// test 2 Write: the first one should setup the connection
+	for i := 0; i < 2; i++ {
+		msg := []byte("some data")
+		n, err := w.Write(msg)
+		require.NoError(t, err)
+		assert.Equal(t, len(msg), n)
+
+		if conn == nil {
+			conn, err = listener.Accept()
+			require.NoError(t, err)
+		}
+
+		var l uint32
+		binary.Read(conn, binary.LittleEndian, &l)
+		assert.Equal(t, uint32(len(msg)), l)
+
+		buffer := make([]byte, 100)
+		n, err = conn.Read(buffer)
+		require.NoError(t, err)
+		assert.Equal(t, "some data", string(buffer[:n]))
+
+		// Unset connection for the next Read
+		w.unsetConnection()
+		conn = nil
 	}
 }