+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 [//]: # (comment: Don't forget to update statsd/telemetry.go:clientVersionTelemetryTag when releasing a new version)
 
+# 5.4.0 / xxxx-xx-xx
+
+- [FEATURE] Add `WithMaxSamplesPerContext()` option to limit the number of samples per context. See [#292][].
+- [BUGFIX] Fix the `rate` of distributions and histograms when using client side aggregation. See [#283][].
+
 # 5.3.0 / 2023-03-06
  
 - [FEATURE] Added support for `DD_DOGSTATSD_URL`. You can now use this env var to set the URL to use to connect to DogStatsD. See [#273][]